pyproject.toml
[project]
name = "python-propagate"
version = "0.1.0"
description = "Orbital Propagation Tool"
authors = [
    {name = "Aaron Berkhoff",email = "aaronberkhoff@outlook.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "numpy (>=2.2.1,<3.0.0)",
    "scipy (>=1.15.1,<2.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "astropy",
    "spiceypy (>=6.0.0,<7.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pre-commit = "^4.1.0"
pylint = "^3.3.4"
black = "^25.1.0"
mypy = "^1.14.1"

[tool.black]
include = "src/|tests/"  # Only format these folders
exclude = "migrations/|venv/"
line-length = 88

[tool.mypy]
files = ["src"]
ignore_missing_imports = true
strict = true
